+++ js/main.js
@@ -1,4 +1,4 @@
-import getAdvertisements from '/js/modules/generateAdverisement.js';
+import getAdvertisements from '/js/modules/generate-adverisement.js';
 
 // eslint-disable-next-line no-console
 console.log(getAdvertisements());
